Updates to reading extra data and arbitrary calculatoins
This includes initial stage of providing field names in HDF5 data that are Npart x m (2D arrays).

+++ src/allvars.h
@@ -220,7 +220,7 @@ using namespace NBody;
 #define CALCMAXMASSWEIGHT 16
 #define CALCLOGAVERAGEMASSWEIGHT 17
 #define CALCLOGSTDMASSWEIGHT 18
-typedef float (*ExtraPropFunc)(float, float, float&);
+typedef double (*ExtraPropFunc)(double, double, double&);
 
 //@}
